+++ octodns_bind/__init__.py
@@ -177,12 +177,12 @@ class ZoneFileProvider(RfcPopulate, BaseProvider):
         self._zone_records = {}
 
     def list_zones(self):
-        n = len(self.file_extension) - 1
+        n = len(self.file_extension)
         for filename in sorted(listdir(self.directory)):
             if filename.endswith(self.file_extension):
                 if n > 0:
                     filename = filename[:-n]
-                yield filename
+                yield f'{filename}.'
 
     def _load_zone_file(self, zone_name, target):
         if target: